使用 clang 20.1.5; 清理aspnet

+++ build_xanmod_docker.sh
@@ -6,16 +6,18 @@ SAVE_DIR=${PWD}
 [ -e "${GITHUB_WORKSPACE}/runner/01_nodoc" ] && cp -a ${GITHUB_WORKSPACE}/runner/01_nodoc /etc/dpkg/dpkg.cfg.d/ || true
 [ -e "runner/01_nodoc" ] && cp -a runner/01_nodoc /etc/dpkg/dpkg.cfg.d/ || true
 
-export PATH="/opt/llvm19_krl/llvm-x86_64/bin/:/opt/llvm19_krl/llvm-19.1.4-x86_64/bin/:/opt/llvm19_krl/llvm-19.1.7-x86_64/bin/:$PATH"
+export LLVMVER="20.1.5"
+export PATH="/opt/llvm19_krl/llvm-x86_64/bin/:/opt/llvm19_krl/llvm-19.1.4-x86_64/bin/:/opt/llvm19_krl/llvm-19.1.7-x86_64/bin/:/opt/llvm19_krl/llvm-${LLVMVER}-x86_64/bin/:$PATH"
 
 # avoid redownload
 [ -e /opt/llvm19_krl/llvm-x86_64/bin/clang ] && exit 0
+[ -e /opt/llvm19_krl/llvm-${LLVMVER}-x86_64/bin/clang ] && exit 0
 [ -e /opt/llvm19_krl/llvm-19.1.4-x86_64/bin/clang ] && exit 0
 [ -e /opt/llvm19_krl/llvm-19.1.7-x86_64/bin/clang ] && exit 0
 
 # config apt llvm
 sed -i '/llvm-toolchain/d' /etc/apt/sources.list
-echo "deb http://apt.llvm.org/jammy/ llvm-toolchain-jammy-19 main" >> /etc/apt/sources.list
+echo "deb http://apt.llvm.org/jammy/ llvm-toolchain-jammy-20 main" >> /etc/apt/sources.list
 
 mkdir -p /etc/apt/trusted.gpg.d/ || true
 curl -L https://apt.llvm.org/llvm-snapshot.gpg.key -o /etc/apt/trusted.gpg.d/apt.llvm.org.asc
@@ -34,26 +36,26 @@ apt install -y --no-install-suggests --no-install-recommends dosfstools xorriso 
   gettext git git-core gperf gzip haveged intltool jq libc6-dev-i386 libelf-dev libfuse-dev libglib2.0-dev libgmp3-dev libltdl-dev libmpc-dev \
   libmpfr-dev libncurses5-dev libncursesw5-dev libpython3-dev libreadline-dev libssl-dev libtool libz-dev lrzsz mkisofs msmtp nano ninja-build \
   p7zip p7zip-full patch pigz pkgconf python3 python3-pip python3-pyelftools python3-setuptools qemu-utils rsync scons squashfs-tools swig \
-  tar uglifyjs unzip upx upx-ucl vim wget xmlto xsltproc xxd xz-utils yasm zip zlib1g-dev zstd liblzma-dev libpam0g-dev pahole dwarves llvm-19 clang-19
+  tar uglifyjs unzip upx upx-ucl vim wget xmlto xsltproc xxd xz-utils yasm zip zlib1g-dev zstd liblzma-dev libpam0g-dev pahole dwarves llvm-20 clang-20
 
 apt-get clean
 which llc || true
-sudo -E ln -s /usr/bin/llc-19 /usr/bin/llc || true
-sudo -E ln -s /usr/bin/clang-19 /usr/bin/clang  || true
+sudo -E ln -s /usr/bin/llc-20 /usr/bin/llc || true
+sudo -E ln -s /usr/bin/clang-20 /usr/bin/clang  || true
 ls -al /usr/bin/llc* || true
 
 # llvm pgo version
-curl -L https://mirrors.edge.kernel.org/pub/tools/llvm/files/llvm-19.1.7-x86_64.tar.xz -o /dev/shm/llvm19.tar.xz
+curl -L https://mirrors.edge.kernel.org/pub/tools/llvm/files/llvm-${LLVMVER}-x86_64.tar.xz -o /dev/shm/llvm19.tar.xz
 
 sudo -E rm -rf /opt/llvm19_krl || true
 mkdir -p /opt/llvm19_krl || true
 
 tar -C /opt/llvm19_krl -xJf /dev/shm/llvm19.tar.xz
 rm /dev/shm/llvm19.tar.xz || true
-ls -al /opt/llvm19_krl/llvm-19.1.7-x86_64/bin || true
+ls -al /opt/llvm19_krl/llvm-${LLVMVER}-x86_64/bin || true
 
 sudo -E rm /opt/llvm19_krl/llvm-x86_64 || true
-sudo -E ln -s /opt/llvm19_krl/llvm-19.1.7-x86_64 /opt/llvm19_krl/llvm-x86_64 || true
+sudo -E ln -s /opt/llvm19_krl/llvm-${LLVMVER}-x86_64 /opt/llvm19_krl/llvm-x86_64 || true
 ls -al /opt/llvm19_krl/ || true
 
 cd $SAVE_DIR
